Use gulp-hub in project type: static-site
