+++ src/static-site/index.js
@@ -55,6 +55,12 @@ function getStaticFiles (props) {
     'pages/components/stylesheets.ejs',
     'pages/views/index.ejs',
     'styles/main.less',
+    'tasks/build.js',
+    'tasks/clean.js',
+    'tasks/js.js',
+    'tasks/pages.js',
+    'tasks/styles.js',
+    'tasks/watch.js',
     'test/ui/index.js',
     'gulpfile.babel.js',
     'webpack.dev.js',
