+++ src/static-site/templates/tasks/clean.js
@@ -1,0 +1,6 @@
+import gulp from 'gulp'
+import del from 'del'
+
+gulp.task('clean', () => {
+  return del('../dist/**/*')
+})
